Refactored file system
