+++ .idea/misc.xml
@@ -16,6 +16,9 @@
         <entry key="../../../../layout/compose-model-1661101819914.xml" value="0.75" />
         <entry key="../../../../layout/compose-model-1661102920407.xml" value="0.75" />
         <entry key="../../../../layout/compose-model-1661103006396.xml" value="0.33" />
+        <entry key="../../../../layout/compose-model-1661151719216.xml" value="0.33" />
+        <entry key="../../../../layout/compose-model-1661151845206.xml" value="0.75" />
+        <entry key="../../../../layout/compose-model-1661173716742.xml" value="0.25" />
         <entry key="app/src/main/res/drawable/ic_pin.xml" value="0.2385" />
         <entry key="app/src/main/res/font/a.xml" value="0.37948717948717947" />
       </map>
