+++ app/src/main/java/com/otaman/weather/ui/MainActivity.kt
@@ -1,0 +1,26 @@
+package com.otaman.weather.ui
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import androidx.compose.foundation.layout.*
+import androidx.compose.material.*
+import androidx.compose.ui.Modifier
+import com.otaman.weather.ui.theme.WeatherTheme
+
+class MainActivity : ComponentActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContent {
+            WeatherTheme {
+                // A surface container using the 'background' color from the theme
+                Surface(
+                    modifier = Modifier.fillMaxSize(),
+                    color = MaterialTheme.colors.background
+                ) {
+                    WeatherApp()
+                }
+            }
+        }
+    }
+}
